+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**Version: 0.0.70 (BETA)**
+
+* Fixed some GUI bugs with the Instant Furnace. 
+* Instant Furnace now has recipe.
+
+-------------------------------------------------------------------
 **Version: 0.0.69 (BETA)**
 
 * Added new block. Instant Furnace. Instantly smelts using RF.